move debug dump test to debug dump integration test
