+++ src/server/debug/server/BUILD.bazel
@@ -64,18 +64,15 @@ go_library(
     ],
 )
 
+# gazelle:go_test file
 go_test(
     name = "server_test",
     size = "small",
-    srcs = [
-        "log_test.go",
-        "server_test.go",
-    ],
+    srcs = ["server_test.go"],
     embed = [":server"],
     deps = [
         "//src/debug",
         "//src/internal/errors",
-        "//src/internal/log",
         "//src/internal/lokiutil",
         "//src/internal/lokiutil/client",
         "//src/internal/pachconfig",
@@ -86,6 +83,39 @@ go_test(
         "//src/pps",
         "@com_github_google_go_cmp//cmp",
         "@in_gopkg_yaml_v3//:yaml_v3",
+        "@io_k8s_api//core/v1:core",
+        "@io_k8s_apimachinery//pkg/apis/meta/v1:meta",
+        "@io_k8s_client_go//kubernetes",
+        "@io_k8s_client_go//kubernetes/fake",
+        "@org_golang_google_protobuf//testing/protocmp",
+    ],
+)
+
+go_test(
+    name = "integration_test",
+    srcs = ["integration_test.go"],
+    deps = [
+        "//src/debug",
+        "//src/internal/client",
+        "//src/internal/errors",
+        "//src/internal/pachd",
+        "//src/internal/require",
+        "//src/internal/tarutil",
+        "//src/internal/testutil",
+        "//src/pfs",
+    ],
+)
+
+go_test(
+    name = "log_test",
+    srcs = ["log_test.go"],
+    embed = [":server"],
+    deps = [
+        "//src/debug",
+        "//src/internal/log",
+        "//src/internal/pachconfig",
+        "//src/internal/pctx",
+        "@com_github_google_go_cmp//cmp",
         "@io_k8s_api//core/v1:core",
         "@io_k8s_apimachinery//pkg/apis/meta/v1:meta",
         "@io_k8s_client_go//kubernetes",
